libwg: make wgTurnOnWithProxyFd() available on iOS
